# propertychain/docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: PropertyChain API
  description: |
    RESTful API for PropertyChain - Real Estate Tokenization Platform
    
    ## Authentication
    Most endpoints require authentication via JWT tokens obtained through the `/auth/login` endpoint.
    Include the token in the `Authorization` header as `Bearer <token>`.
    
    ## Rate Limiting
    API requests are rate-limited to 100 requests per minute for authenticated users.
    
    ## Base URL
    Production: `https://api.propertychain.com/v1`
    Staging: `https://staging-api.propertychain.com/v1`
  version: 1.0.0
  contact:
    name: PropertyChain Support
    email: support@propertychain.com
    url: https://propertychain.com/support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.propertychain.com/v1
    description: Production server
  - url: https://staging-api.propertychain.com/v1
    description: Staging server
  - url: http://localhost:3000/api
    description: Development server
...
paths:
...
  /properties:
    get:
      tags: [Properties]
      summary: List all properties
      operationId: listProperties
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
            maximum: 100
        - name: sort
          in: query
          schema:
            type: string
            enum: [price_asc, price_desc, roi_desc, recent]
            default: recent
        - name: location
          in: query
          schema:
            type: string
          description: Filter by location
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: propertyType
          in: query
          schema:
            type: string
            enum: [apartment, house, commercial, land]
      responses:
        200:
          description: List of properties
          content:
            application/json:
              schema:
                type: object
                properties:
                  properties:
                    type: array
                    items:
                      $ref: '#/components/schemas/Property'
                  pagination:
                    $ref: '#/components/schemas/Pagination'

    post:
      tags: [Properties]
      summary: Create new property listing
      operationId: createProperty
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProperty'
      responses:
        201:
          description: Property created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Property'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Property:
      type: object
      properties:
        id:
          type: string
          example: "prop_123456"
        title:
          type: string
          example: "Modern Downtown Apartment"
        location:
          type: string
          example: "Manhattan, New York"
        price:
          type: number
          example: 500000
        tokenPrice:
          type: number
          example: 100
        totalTokens:
          type: integer
          example: 5000
        availableTokens:
          type: integer
          example: 2500
        expectedROI:
          type: number
          example: 7.5
        propertyType:
          type: string
          enum: [apartment, house, commercial, land]
        images:
          type: array
          items:
            type: string
            format: uri
...
    CreateProperty:
      type: object
      required: [title, location, price, tokenPrice, totalTokens, propertyType]
      properties:
        title:
          type: string
        description:
          type: string
        location:
          type: string
        address:
          $ref: '#/components/schemas/Address'
        price:
          type: number
        tokenPrice:
          type: number
        totalTokens:
          type: integer
        expectedROI:
          type: number
        propertyType:
          type: string
          enum: [apartment, house, commercial, land]
        features:
          type: array
          items:
            type: string
        bedrooms:
          type: integer
        bathrooms:
          type: integer
        squareFeet:
          type: integer
        yearBuilt:
          type: integer
...
    Address:
      type: object
      properties:
        street:
          type: string
        city:
          type: string
        state:
          type: string
        zipCode:
          type: string
        country:
          type: string
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
  responses:
...
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...